Move aperture to its own module
