+++ opossum_core/examples/ray_propagation.rs
@@ -1,7 +1,7 @@
 use opossum_core::{
     OpmDocument,
     analyzers::{AnalyzerType, RayTraceConfig},
-    aperture::{Aperture, CircleConfig},
+    apertures::{Aperture, CircleConfig},
     error::OpmResult,
     joule, millimeter,
     nodes::{Lens, NodeGroup, RayPropagationVisualizer, round_collimated_ray_source},
