+++ opossum_core/examples/workshop_03_kepler_real_lenses_wavefront.rs
@@ -1,7 +1,7 @@
 use opossum_core::{
     OpmDocument,
     analyzers::{AnalyzerType, RayTraceConfig},
-    aperture::{Aperture, CircleConfig},
+    apertures::{Aperture, CircleConfig},
     energy_distributions::UniformDist,
     error::OpmResult,
     joule,
