+++ opossum_core/src/apertures/mod.rs
